moved add_channel function to mp_utils
