+++ micro_dl/utils/meta_utils.py
@@ -42,50 +42,6 @@ def write_meta_field(position: ngff.Position, metadata, field_name, subfield_nam
         position.zattrs[field_name] = field_metadata
 
 
-def add_channel(
-    position: ngff.Position,
-    new_channel_array,
-    new_channel_name,
-    overwrite_ok=False,
-):
-    """
-    Adds a channels to the data array at position "position". Note that there is
-    only one 'tracked' data array in current HCS spec at each position. Also
-    updates the 'omero' channel-tracking metadata to track the new channel.
-
-    The 'new_channel_array' must match the dimensions of the current array in
-    all positions but the channel position (1) and have the same datatype
-
-    Note: to maintain HCS compatibility of the zarr store, all positions (wells)
-    must maintain arrays with congruent channels. That is, if you add a channel
-    to one position of an HCS compatible zarr store, an additional channel must
-    be added to every position in that store to maintain HCS compatibility.
-
-    :param Position zarr_dir: NGFF position node object
-    :param np.ndarray new_channel_array: array to add as new channel with matching
-                            dimensions (except channel dim) and dtype
-    :param str new_channel_name: name of new channel
-    :param bool overwrite_ok: if true, if a channel with the same name as
-                            'new_channel_name' is found, will overwrite
-    """
-    assert len(new_channel_array.shape) == len(position.data.shape) - 1, (
-        "New channel array must match all dimensions of the position array, "
-        "except in the inferred channel dimension: "
-        f"array shape: {position.data.shape}"
-        f", expected channel shape: {(position.data.shape[0], ) + position.data.shape[2:]}"
-        f", received channel shape: {new_channel_array.shape}"
-    )
-    # determine whether to overwrite or append
-    if new_channel_name in position.channel_names and overwrite_ok:
-        new_channel_index = list(position.channel_names).index(new_channel_name)
-    else:
-        new_channel_index = len(position.channel_names)
-        position.append_channel(new_channel_name, resize_arrays=True)
-
-    # replace or append channel
-    position["0"][:, new_channel_index] = new_channel_array
-
-
 def generate_normalization_metadata(
     zarr_dir,
     num_workers=4,
